Update and fix the line tracer code. It works

+++ src/Function.h
@@ -1,4 +1,3 @@
-
 void readSensors();
 void moveMotor(int leftSpeed, int rightSpeed, int duration);
 void followLine();
@@ -8,6 +7,6 @@ void Trace_to_Left_Junction();
 void turnLeft();
 void turnRight();
 int calculateError();
-float computePID(int error);
+float computePID(float error);
 void setLinetrackingBaseSpeed(int setBaseSpeed, int setMinSpeed, int setMaxSpeed);
 void EnergyDetection();
